Extracted database into its own directory
