+++ transaction_service/src/commands/display_sum_cmd.py
@@ -1,9 +1,11 @@
 import sys
 import pymongo
 
-from .db_log import dbLog
 
-from ..database.database import Database
+sys.path.append('../../')
+
+from database.src.database import Database
+from database.src.db_log import dbLog
 
 ACCOUNTS_COLLECT = "accounts"
 TRANSACT_COLLECT = "transactions"
