+++ transaction_service/src/commands/sell_trigger_cmd.py
@@ -2,10 +2,10 @@ import sys
 
 import pymongo
 
-from .db_log import dbLog
+sys.path.append('../../')
 
-from ..database.database import Database
-
+from database.src.database import Database
+from database.src.db_log import dbLog
 
 ACCOUNTS_COLLECT = "accounts"
 TRIGGER_COLLECT = "triggers"
