+++ transaction_service/src/transaction_server.py
@@ -17,134 +17,11 @@ from .commands.sell_trigger_cmd import SetSellAmtCmd, SetSellTriggerCmd, CancelS
 from .commands.display_sum_cmd import DisplaySumCmd
 from .commands.dumplog_cmd import DumplogCmd
 
-
-from .database.database import Database
-from threading import Lock
-from threading import Thread
-
-# Default server values, but can be inputted by user on class initialization
-SERV_PORT       = 55000
-SERV_HOST_NAME  = '127.0.0.1'
-
-
-ACCOUNTS_COLLECT = "accounts"
-TRANSACT_COLLECT = "transactions"
-TRIGGER_COLLECT = "triggers"
-REMOVE = "REMOVE"
-ADD = "ADD"
-
-PACKET_SIZE      = 1024
+sys.path.append('../../')
+from database.src.database import Database
 
 # Connect to database
 Database.connect()
-
-class transactionServer:
-
-    def __init__(self, port=SERV_PORT, hostname=SERV_HOST_NAME):
-        self.port           = port
-        self.hostname       = hostname
-        self.userProcesses  = {}
-        self.serverRunning  = True
-
-
-    def start(self):
-        """ starts the transaction server up. This includes initializing the socket,
-                binding the socket for receiving, and calling the class listen function"""
-        time_now = time.strftime("%a, %d %b %Y %H:%M:%S", time.localtime())
-        print("-------SERVER STARTED!-------")
-        print("Timestamp: ", time_now)
-        print("Port: " + str(self.hostname))
-        print("Port: " + str(self.port))
-        print()
-
-        self.socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        try:
-            self.socket.bind((self.hostname, self.port))
-
-        except socket.error as err:
-            print(f"ERROR! binding port {self.port} failed with error: {err}")
-            self.socket.close()
-            self.shutdown()
-            sys.exit(1)
-
-        # Call to the listening function to start recieving and decoding incoming packets
-        self.listen()
-
-    def shutdown(self):
-        """ shutsdown the server connection, and socket. """
-        try:
-            self.socket.shutdown(socket.SHUT_RDWR)
-
-            print("~ TRANSACTION SERVER SHUTDOWN")
-            time_now = time.strftime("%a, %d %b %Y %H:%M:%S", time.localtime())
-            print("Timestamp: ", time_now)
-
-            self.serverRunning = False
-
-        except socket.error as err:
-            pass
-
-
-    def listen(self):
-
-        self.socket.listen()
-
-        while self.serverRunning:
-            (conn, address) = self.socket.accept()
-
-            conn.setblocking(False)
-
-            self.handleCmd(conn, address)
-
-        self.shutdown()
-
-    def handleConnection(self, conn, address):
-        """ Listens to the socket for incoming packets.  """
-
-        print("Server Connection Detected from: {}".format(address))
-
-        while self.serverRunning:
-            data = []
-
-            try:
-                data = conn.recv(PACKET_SIZE)
-                if not data:
-                    # Connection has been closed
-                    break
-            except socket.error as err:
-                time.sleep(1)
-                continue
-
-            print("~~ Read Data ~~")
-
-            # Sometimes Data packets are bunched up in the read buffer.
-            #      This mechanism will separate them, and process each
-            strData = str(data).strip("b\"")
-
-            for userReqData in packets[:-1]:
-                self.handleRequest(strData)
-
-        conn.shutdown(socket.SHUT_RDWR)
-        conn.close()
-        print("WARNING! Server Connection {conn} closed")
-
-
-    def handleRequest(self, rawPacket):
-
-
-        userReq = ast.literal_eval(str(rawPacket))
-
-        userCommands[command](userReq, threadContext, time.time())
-
-
-if __name__ == '__main__':
-    host_adr = input("Enter hostname (localhost default): ") or "localhost"
-    port = int(input("Enter port number ({} default): ".format(SERV_PORT)) or SERV_PORT)
-
-    server = transactionServer(port, host_adr)
-    server.start()
-
-
 
 def printCmd(cmdDict):
     """
